+++ .editorconfig
@@ -10,6 +10,3 @@ insert_final_newline = true
 [*.{json,yml}]
 indent_size = 2
 indent_style = space
-
-[*.md]
-trim_trailing_whitespace = false
