+++ .gitattributes
@@ -1,1 +1,2 @@
 * text=auto
+*.js text eol=lf
