+++ .travis.yml
@@ -1,4 +1,5 @@
 sudo: false
 language: node_js
 node_js:
-  - 'stable'
+  - '6'
+  - '4'
